requirements.txt
numpy
pandas
duckdb
docling
langchain
langchain-community
chromadb
langgraph
sentence-transformers
ollama
langchain-ollama
streamlit
pydantic
python-dotenv 
deepeval
ragas
langsmith
transformers
scikit-learn
plotly
fastapi
pydpdf
langgraph
streamlit
requests
python-multipart
python-dotenv